+++ myokit/_system.py
@@ -65,6 +65,14 @@ def system(live_printing=False):
     out.append('== GUI ==')
 
     try:    # pragma: no cover
+        from PyQt6.QtCore import QT_VERSION_STR
+        out.append('PyQt6: ' + QT_VERSION_STR)
+        out.append('  Sip: ' + _module_version('sip'))
+        del QT_VERSION_STR
+    except ImportError:
+        out.append('PyQt6: Not found')
+
+    try:    # pragma: no cover
         from PyQt5.QtCore import QT_VERSION_STR
         out.append('PyQt5: ' + QT_VERSION_STR)
         out.append('  Sip: ' + _module_version('sip'))
@@ -72,18 +80,7 @@ def system(live_printing=False):
     except ImportError:
         out.append('PyQt5: Not found')
 
-    try:    # pragma: no cover
-        from PyQt4.QtCore import QT_VERSION_STR
-        out.append('PyQt4: ' + QT_VERSION_STR)
-        out.append('  Sip: ' + _module_version('sip'))
-        del QT_VERSION_STR
-    except ImportError:
-        out.append('PyQt4: Not found')
-    except RuntimeError:    # pragma: no cover
-        # Happens if PyQt5 was also found
-        out.append('PyQt4: OK')
-
-    out.append('PySide: ' + _module_version('PySide'))
+    out.append('PySide6: ' + _module_version('PySide6'))
     out.append('PySide2: ' + _module_version('PySide2'))
     out.append('')